requirements.txt
cffi==1.16.0
cryptography==41.0.7
distro==1.8.0
pycparser==2.21
pyOpenSSL==23.3.0
